requirements.txt
flask
requests
beautifulsoup4
duckduckgo-search
gunicorn
requests
beautifulsoup4
googlesearch-python
python-whois
dnspython
